Un-export Logger fields
This package was split from https://blitiri.com.ar/p/chasquid/, so we
need to do some minor adjustments for publishing it independently.

This patch un-exports some fields from Logger, which is not obvious are
needed externally, and could complicate things later if we want to
change them once someone is relying on them.

They can always be added back if needed.