Add Go modules support
This patch adds Go modules support, so users can incorporate this
package more easily when using them.

Since the package has no dependencies, the module itself is quite easy.

This also makes the "v1" branch obsolete, as now versioning can be done
properly with the modules system. Documentation is updated to reflect
this.